+++ net.luis.main/application/config/testdynamicds.config.properties
@@ -1,0 +1,4 @@
+driver-class-for-dyn-ds=test.net.jplugin.core.das.route.dynamicds.DynamicDsTestProvider
+default=ds1
+1001=asdf
+1122=asdf
