--- net.luis.main/bin/.gitignore
@@ -1,1 +1,0 @@
-/net/
